pricelistUpload

+++ src/test/java/market/api/model/OrderInputDto.java
@@ -2,10 +2,10 @@ package market.api.model;
 
 public class OrderInputDto {
     private Delivery delivery;
-    private Long itemPriceId;
+    private Integer itemPriceId;
     private Integer quantity;
 
-    public OrderInputDto(Delivery delivery, Long itemPriceId, Integer quantity) {
+    public OrderInputDto(Delivery delivery, Integer itemPriceId, Integer quantity) {
         this.delivery = delivery;
         this.itemPriceId = itemPriceId;
         this.quantity = quantity;
